refactor(wallet): address review on ApprovalController wiring
- Restructure into a per-controller directory
  (`instances/approval-controller/{approval-controller,types}.ts`) so it can
  hold a dedicated type + future utils (review: @matthewwalsh0).
- Extract an `ApprovalControllerInstanceOptions` type and reference it from
  `InstanceSpecificOptions` (review: @matthewwalsh0).
- Add `snap_dialog` to the default rate-limit-exclusion set (superset baseline
  the clients share); the smart-tx status type stays client-injected since its
  string differs per platform (review: @FrederikBolding, @matthewwalsh0).
- Mark the changelog entry BREAKING: the default Wallet now registers
  `ApprovalController:*`, so consumers wiring their own must remove it on
  upgrade to avoid a messenger collision (review: @FrederikBolding).
- Point the CODEOWNERS entry at the new directory.
