+++ packages/wallet/src/initialization/instances/index.ts
@@ -1,3 +1,3 @@
-export { approvalController } from './approval-controller';
+export { approvalController } from './approval-controller/approval-controller';
 export { keyringController } from './keyring-controller';
 export { storageService } from './storage-service';
